+++ src/roadtraffic/__init__.py
@@ -1,6 +1,7 @@
 from . import fintraffic
-from . import utils
+from . import plots
 from . import traffic_data_classes
+from . import utils
 
 # Define __all__
-__all__ = ["fintraffic",  "utils", "traffic_data_classes"]
+__all__ = ["fintraffic", "utils", "traffic_data_classes", "plots"]