+++ src/roadtraffic/traffic_data_classes/__init__.py
@@ -1,5 +1,6 @@
-from . import raw_data
 from . import aggregated_data
 from . import bagged_data
+from . import raw_data
+from . import rolling_data
 
-__all__ = ["raw_data", "aggregated_data", "bagged_data"]
+__all__ = ["raw_data", "aggregated_data", "bagged_data", "rolling_data"]