+++ src/roadtraffic/utils/__init__.py
@@ -1,7 +1,7 @@
 from . import constants
+from . import funcs
 from . import load
 from . import models
 from . import process
-from . import funcs
 
 __all__ = ["process", "models", "constants", "load", "funcs"]